orchestrator/pyproject.toml
[project]
name = "murmur-orchestrator"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "aiosqlite>=0.20",
    "pyjwt>=2.9",
    "bcrypt>=4.2",
    "python-multipart>=0.0.18",
    "httpx>=0.28",
    "sse-starlette>=2.0",
    "email-validator>=2.3.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
